Added a file:// scheme.
